back/requirements.txt
django
djangorestframework
mysqlclient
django-cors-headers
cloudinary
PyJWT==v1.7.1